Move XSTS util classes
