+++ subprojects/xsts/xsts-analysis/src/main/kotlin/hu/bme/mit/theta/xsts/analysis/XstsSerializer.kt
@@ -1,4 +1,4 @@
-/*
+package hu.bme.mit.theta.xsts.analysis/*
  *  Copyright 2024 Budapest University of Technology and Economics
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
